Support configurable CMake generator with Ninja and Make defaults
This change introduces flexibility in selecting the CMake generator by using the CMAKE_GENERATOR environment variable, defaulting to "Ninja". The build script now dynamically chooses between Ninja and Make based on this setting. The NINJA_TOOL variable is updated accordingly to ensure compatibility when Make is selected as the generator.

* Build system
* CMake configuration

** Generated by CodeLite. **

Signed-off-by: Eran Ifrah <eifrah@amazon.com>